Add metadata module
Define the necessary structs and enums for every encryption-related
metadata. Create a container for all these values, the `Metadata`
struct, that can be serialized to a buffer. The serialization uses
Protocol Buffers under the hood, so this module provides conversions to
and from the respective protobuf messages. Finally, provide ample
documentation and tests for each metadata type.