Add AEAD helpers
Add some wrappers over the AEAD functions of the `ring` library. These
wrappers convert arrays of u8 values to the types that `ring` wants for
each of its parameters. They perform some rudimentary checks for
the sizes of the keys and the data buffers, to catch simple user errors.
For decryption errors, they return a blanket `DecryptionError`, since
`ring` does not provide more context for security reasons.

Finally, add some tests and documentation on the requirements of these
wrappers and an example of how to use them.