Add cryptor module
Introduce the notion of cryptors, which are structs that encrypt and
decrypt data buffers, based on user-specified metadata. Provide a
cryptor called `RingCryptor`, called this way because it depends on the
cryptographic primitives of the `ring` crate. Finally, document how this
cryptor can be used to encrypt and decrypt data, and add various tests.