proto: Skip rustfmt on auto-generated file
While auto-generated files should ideally follow the formatting rules of
the language, this cannot always be the case, and the formatting may
drift. In order to ignore auto-generated files when running `rustfmt`,
we have two options:

1. Use the `ignore` directive [1], which as of writing this is not
   available in stable Rust.
2. Use the `#[rustfmt::skip]` attribute on the `mod.rs` file, which is
   a bit uglier but is available in stable Rust.

We decided to go with the second option for now, but once the first
option is stabilized, we should switch to that.

[1]: https://github.com/rust-lang/rustfmt/blob/master/Configurations.md#configuration-file-resolution