cryptors: Add an easy and safe way to derive a key
Add a convenience wrapper around `derive_key_no_alloc` called
`derive_key`, which performs the key buffer allocation itself. Normally
this is dangerous [1], since it leaves the key in the heap and it may
remain there for a while. However, we dampen this risk by using the
`Zeroizing` trait from the `zeroize` [2] crate, which will zero the key
buffer before it's dropped.

[1]: https://crypto.stackexchange.com/questions/9998/should-i-delete-cryptographic-data-from-memory
[2]: https://crates.io/crates/zeroize