cryptors: Derive keys only in specific methods
`RingCryptor` has the following public methods for
encryption/decryption:

* seal_with_passphrase
* seal_with_key
* seal_with_meta
* seal_in_place
* open
* open_with_meta
* open_in_place

In all of these methods, we previously used to perform key derivation.
The main reason was that the user could not do it themselves, but this
has been changed with the introduction of `derive_key`. On the other
hand, the problem with always performing key derivation internally is
that it can be slow, which is especially irritating when the same
passphrase is used successively. In these cases, it's better to allow
the user choose when and how they will derive the keys.

Looking into these methods, it seems that are two where key derivation
must remain:

1. seal_with_passphrase: In this method, we generate the key derivation
   metadata internally, so it's impossible to provide a derived key
   beforehand.
2. open: Callers of this method typically don't have access to the
   metadata struct, and thus can't know the key derivation metadata.

In all of the other methods, there is either no key derivation going on,
or they are more low-level. So, it makes sense to disable key derivation
in these cases, and expect that the user will pass the proper key.

Closes #6