Update Cargo.lock
Run `cargo update` in order to fetch the latest dependencies that match
the Cargo.toml version requirements. As a result, probably due to an
update in our protobuf dependencies, `cargo build` failed with the
following:

    error: use of deprecated struct `protobuf::lazy::Lazy`: Please
    regenerate .rs files from .proto files to use newer APIs

Regenerating the Rust code from the protobuf definition with `cargo
build --features proto-gen` fixes this issue.