ci: Regenerate Rust code when we update dependencies
Updating project dependencies during our tests may lead to an update in
the `protobuf` crate. It seems that such an update does not work well
with already generated Rust code from Protobuf definitions, and our CI
fails with errors like this:

    error[E0425]: cannot find value `VERSION_3_6_0` in crate `protobuf`
      --> src/../proto/metadata.rs:25:49
       |
    25 | const _PROTOBUF_VERSION_CHECK: () = ::protobuf::VERSION_3_6_0;
       |                                                 ^^^^^^^^^^^^^ help: a constant with a similar name exists: `VERSION_3_7_1`
       |
      ::: /home/apyrgio/playground/tindercrypt/target/debug/build/protobuf-81b136a5c49b834c/out/version.rs:7:1
       |
    7  | pub const VERSION_3_7_1: () = ();
       | --------------------------- similarly named constant `VERSION_3_7_1` defined here

See this failed run for instance:
https://github.com/apyrgio/tindercrypt/actions/runs/11622120587/job/32367072109#step:5:157

It's probably best to re-generate this Rust code in our CI, just to be
on the safe side.